Renamed namespaces to .jsonschema
