+++ CsCore/CsCoreUnity/Plugins/CsCoreUnity/com/csutil/ui/jsonschema/views/ListEntryView.cs
@@ -1,10 +1,10 @@
 ﻿using com.csutil.model;
-using com.csutil.model.mtvmtv;
+using com.csutil.model.jsonschema;
 using System.Threading.Tasks;
 using UnityEngine;
 using UnityEngine.UI;
 
-namespace com.csutil.ui.mtvmtv {
+namespace com.csutil.ui.jsonschema {
 
     public class ListEntryView : InputFieldView {
